% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cnst}
\alias{get_cnst}
\title{Get Predefined Constant}
\usage{
get_cnst(s)
}
\arguments{
\item{s}{character scalar, name of the constant whoes value is to get.}
}
\description{
Get Predefined Constant
}
