% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cnst}
\alias{get_cnst}
\title{Get Predefined Constant}
\usage{
get_cnst(s)
}
\arguments{
\item{s}{character scalar, name of the constant whose value is to get.}
}
\value{
integer scalar, the value of the constant named by \code{s}
}
\description{
This function returns numerical value of predefined named constants. Constants defined for this package are: \itemize{
\item CV_ADAMS
\item CV_BDF
\item CV_SIMULTANEOUS
\item CV_STAGGERED
\item CV_STAGGERED1
\item CV_SUCCESS
\item R2SUNDIALS_EVENT_HOLD
\item R2SUNDIALS_EVENT_IGNORE
\item R2SUNDIALS_EVENT_STOP
}
All these constants are exported from the package and available for use in callback functions written in R.
}
\examples{
get_cnst("CV_SUCCESS")
# [1] 0
}
